Move SampleIndex to rl_vector::index
